editor-blender/requirements.txt
fake-bpy-module-4.0
gql[all]
dataclass_wizard
